feat: scaffold the state
now, unfortunately this won't do - the select element can't be opened
automatically. I'll have to think about how to do it - maybe drop it
entirely and build the function into links. Though then it would be
unclear how to click on one without activating another slide. Buttons to
switch to next/previous are not an option because they would require a
lot of clicks. I currently think that a sidebar-style toolbar would make
the most sense